Refactor: extract common test cached field preloading code.
