+++ eventbuscachedfield/src/test/groovy/com/byoutline/eventbuscachedfield/EventBusCachedFieldSpec.groovy
@@ -3,6 +3,7 @@ package com.byoutline.eventbuscachedfield
 import com.byoutline.cachedfield.CachedField
 import com.byoutline.cachedfield.FieldState
 import com.byoutline.cachedfield.FieldStateListener
+import com.byoutline.cachedfield.MockCachedFieldLoader
 import com.byoutline.cachedfield.MockFactory
 import com.byoutline.eventcallback.ResponseEvent
 import de.greenrobot.event.EventBus
@@ -25,23 +26,7 @@ class EventBusCachedFieldSpec extends Specification {
     ResponseEvent<Exception> errorEvent
     EventBus bus
 
-    static void postAndWaitUntilFieldStopsLoading(CachedField field) {
-        boolean duringValueLoad = true
-        def listener = { FieldState newState ->
-            if (newState == FieldState.NOT_LOADED || newState == FieldState.LOADED) {
-                duringValueLoad = false
-            }
-        } as FieldStateListener
 
-        field.addStateListener(listener)
-        field.postValue()
-        while (duringValueLoad) {
-            sleep 1
-        }
-        field.removeStateListener(listener)
-        // allow thread switch from state listener executor to success listener
-        sleep 8
-    }
 
     def setup() {
         bus = Mock()
@@ -80,7 +65,7 @@ class EventBusCachedFieldSpec extends Specification {
                 .withSuccessEvent(successEvent)
                 .withResponseErrorEvent(errorEvent)
                 .build()
-        postAndWaitUntilFieldStopsLoading(field)
+        MockCachedFieldLoader.postAndWaitUntilFieldStopsLoading(field)
 
         then:
         sC * successEvent.setResponse(value)
@@ -102,7 +87,7 @@ class EventBusCachedFieldSpec extends Specification {
                 .build()
 
         when:
-        postAndWaitUntilFieldStopsLoading(field)
+        MockCachedFieldLoader.postAndWaitUntilFieldStopsLoading(field)
 
         then:
         1 * bus.post(expEvent)
@@ -121,7 +106,7 @@ class EventBusCachedFieldSpec extends Specification {
                 .build()
 
         when:
-        postAndWaitUntilFieldStopsLoading(field)
+        MockCachedFieldLoader.postAndWaitUntilFieldStopsLoading(field)
 
         then:
         1 * customBus.post(_)
