+++ ottocachedfield/src/test/groovy/com/byoutline/ottocachedfield/PostingToBusCachedFieldWithArgSpec.groovy
@@ -3,6 +3,7 @@ package com.byoutline.ottocachedfield
 import com.byoutline.cachedfield.CachedFieldWithArg
 import com.byoutline.cachedfield.FieldState
 import com.byoutline.cachedfield.FieldStateListener
+import com.byoutline.cachedfield.MockCachedFieldLoader
 import com.byoutline.cachedfield.MockFactory
 import com.byoutline.ibuscachedfield.events.ResponseEventWithArg
 import com.squareup.otto.Bus
@@ -26,32 +27,6 @@ class PostingToBusCachedFieldWithArgSpec extends Specification {
     ResponseEventWithArg<Exception, Integer> errorEvent
     Bus bus
 
-    static <ARG_TYPE> void postAndWaitUntilFieldStopsLoading(CachedFieldWithArg<?, ARG_TYPE> field, ARG_TYPE arg) {
-        boolean duringValueLoad = true
-        boolean loadingStarted = false
-        def listener = { FieldState newState ->
-            switch (newState) {
-                case FieldState.NOT_LOADED:
-                    if (loadingStarted) duringValueLoad = false
-                    break
-                case FieldState.CURRENTLY_LOADING:
-                    loadingStarted = true
-                    break
-                case FieldState.LOADED:
-                    duringValueLoad = false
-                    break
-            }
-        } as FieldStateListener
-
-        field.addStateListener(listener)
-        field.postValue(arg)
-        while (duringValueLoad) {
-            sleep 1
-        }
-        field.removeStateListener(listener)
-        sleep 8  // wait for event to be posted
-    }
-
     def setup() {
         bus = Mock()
         successEvent = Mock()
@@ -69,7 +44,7 @@ class PostingToBusCachedFieldWithArgSpec extends Specification {
                 .withResponseErrorEvent(errorEvent)
                 .build()
         when:
-        postAndWaitUntilFieldStopsLoading(field, arg)
+        MockCachedFieldLoader.postAndWaitUntilFieldStopsLoading(field, arg)
 
         then:
 
@@ -98,7 +73,7 @@ class PostingToBusCachedFieldWithArgSpec extends Specification {
                 .withResponseErrorEvent(errorEvent)
                 .build()
         when:
-        postAndWaitUntilFieldStopsLoading(field, 2)
+        MockCachedFieldLoader.postAndWaitUntilFieldStopsLoading(field, 2)
 
         then:
         errorVal.message == "E2"
@@ -122,7 +97,7 @@ class PostingToBusCachedFieldWithArgSpec extends Specification {
                 .build()
 
         when:
-        postAndWaitUntilFieldStopsLoading(field, 1)
+        MockCachedFieldLoader.postAndWaitUntilFieldStopsLoading(field, 1)
 
         then:
         1 * customBus.post(_)
